backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
pandas==2.2.2
numpy==1.26.4
google-generativeai==0.7.2
sqlparse==0.5.0
python-multipart==0.0.9
pydantic>=2.9.0,<3.0.0
python-dotenv==1.0.1
